Both UDP and QR listeners active; camera and QR works
This is not done yet but it's almost there, and it's time to back up with a commit. HearThisAndroid code for camera and QR code decode is ported and working. The two listeners (UDP, QR) are both active in their separate threads, with UDP going first and QR activating only if UDP doesn't see an advert within 20 seconds.
Next is to hook up the QR decode data to the book request logic; this will require a little refactoring.