First book transfer via QR code instead of UDP broadcast!
Incorporated the other (QR) listener into the refactored mechanism of listening for an advert, then responding to it with a book request. Added simple means for NewBookListenerService to get the decoded QR data from SyncActivity. Improved the logic in which UDP-listener gets first opportunity to get an advert, but if it doesn't within a specified time then QR scan is enabled.
The code is not cleaned up, and there are still a few TODOs, but am committing this now for safety.