Add successful QR scan screen closure
QR scan screen now closes after providing the decoded QR data. We know the scan is complete when the camera shuts down, so at that point SyncActivity goes into a polling loop waiting for NewBookListenerService to set a flag indicating that the data has been recieved. SyncActivity then knows it is safe to close which it does by calling finish().
This code is not yet cleaned up. To enable QR testing, it also still has UDP broadcasts temporarily disabled ('willIgnore' is hardcoded false).