Remove textbox code that simulated QR code content
Now that QR code advert processing is in place this code is no longer needed. Remove it as well as a few other things that I noticed along the way.
This code builds and runs and successfully transfers a book, using UDP adverts, with both Samsung and BQ tablets. Next is to refocus on QR code processing and fix the remaining wrinkles there.