First time QR works with both tablets
Previously it worked with the Samsung but never with the BQ tablet. For some reason the BQ tablet would always fail to start QR subsystem with the comment that maybe the flag 'FLAG_ACTIVITY_NEW_TASK' is needed. I added it, and it works now.
This code has commented-out stuff from a fix attempt that tried to mimic more how HTA starts SyncActivity from the same file. It sort of worked, but also had a fatal error due to a null object reference. That is why it is now commented out.