Add timeout to TCP socket for book request
Last week I ran into a scenario where the TCP socket connection attempt with Desktop never completed, and Reader got locked up. An oversight that I never had a timeout on it; add one now with a 3-second value.
This configuration of Reader files is what I will start with for the first attempt at running the 7 (so far) test cases.