Remove quotes from Reader's SSID
WifiInfo contains the Reader's SSID but with quotes around it. Copying that element into a String retains the quotes. We can't have that because (a) it will create a false negative comparison with Desktop's quote-less SSID even it both Reader and Desktop are on the same network, and (b) quote char is invalid for an SSID. So remove quotes.